fix(design-system): drop DS- prefix; replace webapp Button story with DS Button
- Rename DS-Button.stories.tsx → Button.stories.tsx (replaces the old
  webapp Button story, which will be obsolete after NAN-5946)
- Rename DS-Icon.stories.tsx → Icon.stories.tsx (no conflict existed)
