--- packages/design-system/stories/DS-Button.stories.tsx
@@ -1,109 +1,0 @@
-import { ChevronRight, Plus } from 'lucide-react';
-
-import { Button, IconButton } from '../src/components/ui/button';
-
-import type { Meta, StoryObj } from '@storybook/react';
-
-const meta: Meta = {
-    title: 'Design System/Components/Button',
-    parameters: { layout: 'padded' }
-};
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-const VARIANTS = ['primary', 'secondary', 'outline', 'ghost', 'danger', 'link-danger'] as const;
-const ICON_VARIANTS = ['primary', 'secondary', 'outline', 'ghost', 'danger'] as const;
-const SIZES = ['xs', 'sm', 'md', 'lg'] as const;
-
-export const AllVariants: Story = {
-    name: 'All variants',
-    render: () => (
-        <div style={{ display: 'flex', flexDirection: 'column', gap: 'var(--ds-space-10)' }}>
-            {VARIANTS.map((variant) => (
-                <div key={variant} style={{ display: 'flex', alignItems: 'center', gap: 'var(--ds-space-6)', flexWrap: 'wrap' }}>
-                    <span style={{ fontSize: 'var(--ds-typography-font-size-xs)', color: 'var(--text-secondary)', width: '7rem', flexShrink: 0 }}>
-                        {variant}
-                    </span>
-                    <Button variant={variant}>Default</Button>
-                    <Button variant={variant} disabled>
-                        Disabled
-                    </Button>
-                    <Button variant={variant} loading>
-                        Loading
-                    </Button>
-                    <Button variant={variant} leadingIcon={<Plus size={14} />}>
-                        With icon
-                    </Button>
-                    <Button variant={variant} trailingIcon={<ChevronRight size={14} />}>
-                        Trailing
-                    </Button>
-                </div>
-            ))}
-        </div>
-    )
-};
-
-export const AllSizes: Story = {
-    name: 'All sizes',
-    render: () => (
-        <div style={{ display: 'flex', flexDirection: 'column', gap: 'var(--ds-space-10)' }}>
-            {SIZES.map((size) => (
-                <div key={size} style={{ display: 'flex', alignItems: 'center', gap: 'var(--ds-space-6)' }}>
-                    <span style={{ fontSize: 'var(--ds-typography-font-size-xs)', color: 'var(--text-secondary)', width: '3rem', flexShrink: 0 }}>{size}</span>
-                    <Button size={size}>Button</Button>
-                    <Button size={size} variant="outline">
-                        Outline
-                    </Button>
-                    <Button size={size} leadingIcon={<Plus size={14} />}>
-                        With icon
-                    </Button>
-                </div>
-            ))}
-        </div>
-    )
-};
-
-export const IconButtonVariants: Story = {
-    name: 'IconButton — all variants',
-    render: () => (
-        <div style={{ display: 'flex', flexDirection: 'column', gap: 'var(--ds-space-10)' }}>
-            {ICON_VARIANTS.map((variant) => (
-                <div key={variant} style={{ display: 'flex', alignItems: 'center', gap: 'var(--ds-space-6)', flexWrap: 'wrap' }}>
-                    <span style={{ fontSize: 'var(--ds-typography-font-size-xs)', color: 'var(--text-secondary)', width: '5rem', flexShrink: 0 }}>
-                        {variant}
-                    </span>
-                    <IconButton variant={variant} label={`${variant} default`}>
-                        <Plus />
-                    </IconButton>
-                    <IconButton variant={variant} disabled label={`${variant} disabled`}>
-                        <Plus />
-                    </IconButton>
-                    <IconButton variant={variant} loading label={`${variant} loading`} />
-                </div>
-            ))}
-        </div>
-    )
-};
-
-export const IconButtonSizes: Story = {
-    name: 'IconButton — all sizes',
-    render: () => (
-        <div style={{ display: 'flex', flexDirection: 'column', gap: 'var(--ds-space-10)' }}>
-            {SIZES.map((size) => (
-                <div key={size} style={{ display: 'flex', alignItems: 'center', gap: 'var(--ds-space-6)' }}>
-                    <span style={{ fontSize: 'var(--ds-typography-font-size-xs)', color: 'var(--text-secondary)', width: '3rem', flexShrink: 0 }}>{size}</span>
-                    <IconButton size={size} label={`${size} primary`}>
-                        <Plus />
-                    </IconButton>
-                    <IconButton size={size} variant="outline" label={`${size} outline`}>
-                        <Plus />
-                    </IconButton>
-                    <IconButton size={size} variant="ghost" label={`${size} ghost`}>
-                        <Plus />
-                    </IconButton>
-                </div>
-            ))}
-        </div>
-    )
-};
